PlayerCmd hook (experimental)
Adds a PlayerCmd hook to Lua. Works like botcmd but allows to modify a regular player's cmd input.
This hook runs after the player's cmd has been built, so you can read the cmd values for the current frame and change them at will,
In its current implementation, it is clientsided and can even be used from this EXE and not desynch in netgames.

The hook is set up like this:
addHook("PlayerCmd", function(player, cmd) ...

player_t should however not be modified in ways that may impact netgames such that they may desynch as this hook runs into a clientsided fashion. So be careful.